refactor: CurrenciesList component
