+++ src/components/CurrenciesList/index.test.tsx
@@ -1,15 +1,17 @@
 import { BrowserRouter } from 'react-router-dom';
 import { render, screen } from '@testing-library/react';
 
-import { DEFAULT_QUOTES } from '@/db/defaultCurrencies';
+import { DEFAULT_CURRENCIES } from '@/constants/defaultCurrencies';
 
 import '@testing-library/jest-dom';
 
-import CurrenciesList from './CurrenciesList';
+import { CurrenciesList } from '.';
 
 describe('Testing CurrenciesList conponent', () => {
   beforeEach(() =>
-    render(<CurrenciesList title="Test list" data={DEFAULT_QUOTES} />, { wrapper: BrowserRouter }),
+    render(<CurrenciesList title="Test list" data={DEFAULT_CURRENCIES} />, {
+      wrapper: BrowserRouter,
+    }),
   );
 
   it('Should be displayed', () => {
@@ -19,6 +21,6 @@ describe('Testing CurrenciesList conponent', () => {
     const cards = screen.getAllByTestId('currency-card');
 
     expect(title).toHaveTextContent('Test list');
-    expect(cards.length).toBe(DEFAULT_QUOTES.length);
+    expect(cards.length).toBe(DEFAULT_CURRENCIES.length);
   });
 });
