+++ src/components/CurrenciesList/styles.module.scss
